rename constants directory
