+++ src/inchPrice.ts
@@ -5,7 +5,7 @@ import { BigNumber, ethers } from "ethers";
 import axios from "axios";
 import { chainId, protocols, initialAmount, diffAmount } from "./config";
 import { IRoute } from "./interfaces/main";
-import { ERC20Token, IToken } from "./constrants/addresses";
+import { ERC20Token, IToken } from "./constants/addresses";
 import { replaceTokenAddress } from "./utils";
 
 /**
